% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary.R
\name{datasummary}
\alias{datasummary}
\title{Data Summary}
\usage{
datasummary(type)
}
\arguments{
\item{type}{The kind of output given. The possible outputs are \code{count}, \code{timeline},
\code{proportion}, \code{comparison}. \code{count} displays the frequency of the different
variables within the dataset each year. \code{timeline} creates a line graph of the change
in the number of graduation merit types that Jewish people from Williams receives each year.
\code{proportion} creates a line graph of the change in proportion in the graduation types
that Jewish people receive each year. \code{comparision} does the same thing as proportion,
but also includes the proportion of non-jewish people so comparisons can be made.}
}
\value{
Desired infomation from the kind of output specified by \code{type}.
}
\description{
Uses the data frame generated by the readnames function and creates
graphics associated with the data.
}
