% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readnames.R
\name{readnames}
\alias{readnames}
\title{readnames}
\usage{
readnames(datafile, year)
}
\arguments{
\item{datafile}{the textfile that contains a list of the information of the
the students in that year}

\item{year}{takes in the graduation year of the year of students}
}
\value{
a matrix of the combined years with the data extracted from each person

a dataframe of the datafile in a formatted way
}
\description{
Takes in a datafile of names of Williams college graduates, and then
returns a dataset of the first name, last name, what level of honors,
the major. As well this function tests the last name of each person
to check if the person is of Jewish descent or not.
}
\examples{
year2014 <- readnames("~/names/inst/extdata/names2014.txt", "2014")

}
