% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentdata.R
\docType{data}
\name{studentdata}
\alias{studentdata}
\title{Student Data from 2003 to 2015}
\format{A dataframe with 8 variables
\itemize{
   \item name = The name of the person.
   \item year = The graduation year of the person.
   \item major = The major of the person at college.
   \item honors = The type of honors received with degree.
   \item PhiBetaKappa = TRUE for in PhiBetaKappa. FALSE for not in.
   \item SigmaXI = TRUE for in SigmaXI. FALSE for not in.
   \item degree = The type of degree and merit level graduated with.
   \item ethnicity = The ethnicity of the person. In this package only
         Jewish is tested for.
}}
\usage{
studentdata
}
\description{
This data frame contains information of Williams College students
including majors, honors, and the type of degree received upon
graduation from the class of 2003 to the class of 2015.
}
\keyword{studentdata}
