% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totaldata.R
\name{totaldata}
\alias{totaldata}
\title{totaldata}
\usage{
totaldata()
}
\value{
the complete dataframe, or matrix of the complete dataset with all 8000 students
}
\description{
Runs through the list of students names in all the years, creates the matrixes,
and then Rbinds the matrixes together to create an entire dataset
}
